Refactor monster class and organize packages
