+++ src/com/plantpocalypse/events/ActionGuestRoom.java
@@ -1,4 +1,8 @@
-package com.plantpocalypse;
+package com.plantpocalypse.events;
+
+import com.plantpocalypse.PlantMonster;
+import com.plantpocalypse.Player;
+import com.plantpocalypse.Room;
 
 public class ActionGuestRoom implements Action{
     @Override
